Add many-clients example
This is an example that showcases the use for `accept` and
`borrow_request`. The idea is, that an unknown number of clients will
want to connect to the server. They do so by getting a clone of a
Sender, which they then use to establish a session channel to server.

cc @laumann