Fix the documentation tests to handle destructor bombs
The documentation tests were not correctly closing all of their
channels, and with the newly introduced destructor bombs they fail
horribly. This fixes the tests, but the result turns out to be
exceptionally ugly, and we should definitely try to come up with some
better examples that showcase the `chan_select!` macros without being
horribly verbose about closing channels.

cc: @laumann