start adding multiple offers and such
This is an experiment that introduces arity specialized branching
constructs.

The idea is to add multiple specialized Offer/Choose constructs
according to arity. That way, instead of chaining lots of Offer's and
having to do `sel2().sel2().sel2()`, we'd simply have an `OfferN` with
associated `sel1()`, `sel2()`, ..., and `selN()` methods.

This commit only introduces Offer3 and Choose3 structs, but serves to
give us an idea of how our library would end up looking.