Seal the HasDual trait
By using a trick from the libm crate [1] we can seal the HasDual trait,
making it impossible to implement outside of the session-types crate and
allowing us to mark it stable.

We declare a private submodule and export an empty trait "Sealed" from
it. In the private module we implement the Sealed trait for all exported
protocol types.

This removes the "unsafe" declaration and implementation for all
protocol types.

[1] https://github.com/rust-lang-nursery/libm/blob/master/src/lib.rs