Expand the README
On GitHub, the README.md is often the entry-point for potential new
users. Expand the README with how to get started and a basic example
to let readers get a quick feel for how to use it.

This is a little redundant, because the information is already in the
source code itself, but I figure it would make this library a bit more
approachable.